Start removing unnecessary routing callbacks
