+++ client/my-sites/comments/controller.js
@@ -1,9 +1,6 @@
 import page from '@automattic/calypso-router';
-import { startsWith } from 'lodash';
 import { addQueryArgs, getSiteFragment } from 'calypso/lib/route';
 import CommentView from 'calypso/my-sites/comment/main';
-import { removeNotice } from 'calypso/state/notices/actions';
-import { getNotices } from 'calypso/state/notices/selectors';
 import CommentsManagement from './main';
 
 const mapPendingStatusToUnapproved = ( status ) => ( 'pending' === status ? 'unapproved' : status );
@@ -111,27 +108,5 @@ export const comment = ( context, next ) => {
 		page.redirect( `/comments/all/${ siteFragment }/${ postId }` );
 
 	context.primary = <CommentView { ...{ action, commentId, siteFragment, redirectToPostView } } />;
-	next();
-};
-
-export const redirect = ( { path } ) => {
-	const siteFragment = getSiteFragment( path );
-	if ( siteFragment ) {
-		return page.redirect( `/comments/all/${ siteFragment }` );
-	}
-	return page.redirect( '/comments/all' );
-};
-
-export const clearCommentNotices = ( { store }, next ) => {
-	const nextPath = page.current;
-	if ( ! startsWith( nextPath, '/comments' ) ) {
-		const { getState, dispatch } = store;
-		const notices = getNotices( getState() );
-		notices.forEach( ( { noticeId } ) => {
-			if ( startsWith( noticeId, 'comment-notice' ) ) {
-				dispatch( removeNotice( noticeId ) );
-			}
-		} );
-	}
 	next();
 };
